generators/simple/templates/src/platform/extensionRunner/requirements.txt
requests==2.22.0
flask==1.0.3
pyrapidjson==0.6.3
attrs==19.1.0
inflection==0.3.1
typing-extensions==3.7.4
